requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.10.0
bleach==4.1.0
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flit-core==3.7.1
fonttools==4.30.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.43.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.0
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.8.0
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.3
multivolumefile==0.2.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.10
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pip==21.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.27
protobuf==3.19.4
pure-eval==0.2.2
py7zr==0.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==0.5.0
pycparser==2.21
pycryptodomex==3.14.1
Pygments==2.11.2
pylint==2.6.0
pyparsing==3.0.7
pyppmd==0.17.4
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.2
ray==1.11.0
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
resnet==0.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
setuptools==58.0.4
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
stack-data==0.2.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
terminado==0.13.1
testpath==0.6.0
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2022.2.9
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.14.0
zipfile-deflate64==0.2.0
zipp==3.7.0
